Extract trades table to sub component
